refactor: rename `import_soil_hardness` command to `hardness_load_data`
- Updated command name across the codebase for consistency and clarity.
- Adjusted related tests, views, and README to reflect the change.
