+++ README.md
@@ -143,7 +143,7 @@ python manage.py loaddata .\ai_agent\fixtures\rag_material.json
 
 ```
 python manage.py runserver
-python manage.py import_soil_hardness /path/to/folder
+python manage.py hardness_load_data /path/to/folder
 ```
 
 ## よくつかうメンテナンスコマンド
