+++ soil_analysis/management/commands/hardness_load_data.py
